+++ src/com/kaylerrenslow/armaDialogCreator/data/xml/ResourceRegistryXmlWriter.java
@@ -1,8 +1,8 @@
 package com.kaylerrenslow.armaDialogCreator.data.xml;
 
 import com.kaylerrenslow.armaDialogCreator.data.ExternalResource;
-import com.kaylerrenslow.armaDialogCreator.data.GlobalResourceRegistry;
 import com.kaylerrenslow.armaDialogCreator.data.ResourceRegistry;
+import com.kaylerrenslow.armaDialogCreator.data.WorkspaceResourceRegistry;
 import com.kaylerrenslow.armaDialogCreator.util.KeyValueString;
 import org.jetbrains.annotations.NotNull;
 
@@ -14,17 +14,17 @@ import java.io.IOException;
 public class ResourceRegistryXmlWriter {
 	protected ResourceRegistry resourceRegistry;
 
-	public static class GlobalResourceRegistryXmlWriter extends ResourceRegistryXmlWriter {
-		public GlobalResourceRegistryXmlWriter() {
-			super(GlobalResourceRegistry.getInstance());
+	public static class WorkspaceResourceRegistryXmlWriter extends ResourceRegistryXmlWriter {
+		public WorkspaceResourceRegistryXmlWriter() {
+			super(WorkspaceResourceRegistry.getInstance());
 		}
 
-		private GlobalResourceRegistryXmlWriter(@NotNull ResourceRegistry resourceRegistry) {
+		private WorkspaceResourceRegistryXmlWriter(@NotNull ResourceRegistry resourceRegistry) {
 			super(resourceRegistry);
 		}
 
 		@NotNull
-		private XmlWriterOutputStream getXmlWriterOutputStream(@NotNull GlobalResourceRegistry registry) throws IOException {
+		private XmlWriterOutputStream getXmlWriterOutputStream(@NotNull WorkspaceResourceRegistry registry) throws IOException {
 			if (!registry.getResourcesFile().exists()) {
 				registry.getResourcesFile().createNewFile();
 			}
@@ -42,16 +42,16 @@ public class ResourceRegistryXmlWriter {
 		}
 
 		public static void writeAndClose() throws IOException {
-			writeAndClose(GlobalResourceRegistry.getInstance());
+			writeAndClose(WorkspaceResourceRegistry.getInstance());
 		}
 
 
-		public static void writeAndClose(@NotNull GlobalResourceRegistry registry) throws IOException {
-			new GlobalResourceRegistryXmlWriter(registry).doWriteAndClose();
+		public static void writeAndClose(@NotNull WorkspaceResourceRegistry registry) throws IOException {
+			new WorkspaceResourceRegistryXmlWriter(registry).doWriteAndClose();
 		}
 
 		public void doWriteAndClose() throws IOException {
-			XmlWriterOutputStream fos = getXmlWriterOutputStream((GlobalResourceRegistry) this.resourceRegistry);
+			XmlWriterOutputStream fos = getXmlWriterOutputStream((WorkspaceResourceRegistry) this.resourceRegistry);
 			write(fos);
 			fos.flush();
 			fos.close();