+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/popup/projectInit/CouldNotLoadProjectDialog.java
@@ -2,10 +2,12 @@ package com.kaylerrenslow.armaDialogCreator.gui.main.popup.projectInit;
 
 import com.kaylerrenslow.armaDialogCreator.gui.popup.StageDialog;
 import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
+import com.kaylerrenslow.armaDialogCreator.main.ExceptionHandler;
 import com.kaylerrenslow.armaDialogCreator.main.Lang;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextArea;
 import javafx.scene.layout.VBox;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.ResourceBundle;
 
@@ -13,10 +15,12 @@ import java.util.ResourceBundle;
  @author Kayler
  @since 11/23/2016 */
 public class CouldNotLoadProjectDialog extends StageDialog<VBox> {
-	public CouldNotLoadProjectDialog(Exception e) {
+	public CouldNotLoadProjectDialog(@NotNull Exception e) {
 		super(ArmaDialogCreator.getPrimaryStage(), new VBox(5), null, false, true, false);
 		ResourceBundle bundle = Lang.ApplicationBundle();
-		final TextArea taError = new TextArea(e.getMessage());
+		setTitle(bundle.getString("ProjectLoadError.could_not_load_project"));
+
+		TextArea taError = ExceptionHandler.getExceptionTextArea(e);
 		taError.setEditable(false);
 		myStage.setResizable(false);
 		myRootElement.getChildren().addAll(